requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
python-docx>=1.1.0
python-dotenv>=1.0.1
requests>=2.31.0
tavily-python>=0.5.0
